Updated HordeSplitManager to work with new layout
The Split slider was changed to use whole integers instead of decimal value so changed the manager to handle integers.
Also made the displayed amounts changed to be what both splits would be after splitting (e.g. Horde 1: 5 and Horde 2: 7) making sure it never goes below initial pop which is now used and not a percentage calculation.

Changed the function names in UI_manager so all the related functions are together in the alphabetical list and not all enable/disable are together to aid clarity with unity inspector
Also added toggle functions for each display button for more intuitive use on the action bar.